+++ mmedit/apis/inferencer_utils.py
@@ -18,227 +18,22 @@ from mmengine.infer import BaseInferencer
 from mmengine.registry import init_default_scope
 from mmengine.runner import load_checkpoint
 from mmengine.utils import ProgressBar
-from models.base_models import BaseTranslationModel
 from torch.nn.parallel import scatter
 from torchvision import utils
 
+from mmedit.models.base_models import BaseTranslationModel
 from mmedit.registry import MODELS
-from mmedit.utils import (VIDEO_EXTENSIONS, ConfigType, InputsType, PredType,
-                          ResType, pad_sequence, read_frames, read_image,
-                          register_all_modules, set_random_seed)
+from mmedit.utils import VIDEO_EXTENSIONS
 
 
-class BaseMMEditInferencer(BaseInferencer):
-    """Base inferencer.
-
-    Args:
-        config (str or ConfigType): Model config or the path to it.
-        ckpt (str, optional): Path to the checkpoint.
-        device (str, optional): Device to run inference. If None, the best
-            device will be automatically used.
-        result_out_dir (str): Output directory of images. Defaults to ''.
-    """
-
-    func_kwargs = dict(
-        preprocess=[],
-        forward=[],
-        visualize=['result_out_dir'],
-        postprocess=['get_datasample'])
-    func_order = dict(preprocess=0, forward=1, visualize=2, postprocess=3)
-
-    extra_parameters = dict()
-
-    def __init__(self,
-                 config: Union[ConfigType, str],
-                 ckpt: Optional[str] = None,
-                 device: Optional[str] = None,
-                 extra_parameters: Optional[Dict] = None,
-                 seed: int = 2022,
-                 **kwargs) -> None:
-        # Load config to cfg
-        if device is None:
-            device = torch.device(
-                'cuda' if torch.cuda.is_available() else 'cpu')
-        self.device = device
-        register_all_modules()
-        super().__init__(config, ckpt, device)
-
-        self._init_extra_parameters(extra_parameters)
-        self.base_params = self._dispatch_kwargs(**kwargs)
-        self.seed = seed
-        set_random_seed(self.seed)
-
-    def _init_model(self, cfg: Union[ConfigType, str], ckpt: Optional[str],
-                    device: str) -> None:
-        """Initialize the model with the given config and checkpoint on the
-        specific device."""
-        model = MODELS.build(cfg.model)
-        if ckpt is not None and ckpt != '':
-            ckpt = load_checkpoint(model, ckpt, map_location='cpu')
-        model.cfg = cfg
-        model.to(device)
-        model.eval()
-        return model
-
-    def _init_pipeline(self, cfg: ConfigType) -> Compose:
-        """Initialize the test pipeline."""
-        if 'test_dataloader' in cfg and \
-            'dataset' in cfg.test_dataloader and \
-                'pipeline' in cfg.test_dataloader.dataset:
-            pipeline_cfg = cfg.test_dataloader.dataset.pipeline
-            return Compose(pipeline_cfg)
-        return None
-
-    def _init_extra_parameters(self, extra_parameters: Dict) -> None:
-        """Initialize extra_parameters of each kind of inferencer."""
-        if extra_parameters is not None:
-            for key in self.extra_parameters.keys():
-                if key in extra_parameters.keys():
-                    self.extra_parameters[key] = extra_parameters[key]
-
-    def _update_extra_parameters(self, **kwargs) -> None:
-        """update extra_parameters during run time."""
-        if 'extra_parameters' in kwargs:
-            input_extra_parameters = kwargs['extra_parameters']
-            if input_extra_parameters is not None:
-                for key in self.extra_parameters.keys():
-                    if key in input_extra_parameters.keys():
-                        self.extra_parameters[key] = \
-                            input_extra_parameters[key]
-
-    def _dispatch_kwargs(self, **kwargs) -> Tuple[Dict, Dict, Dict, Dict]:
-        """Dispatch kwargs to preprocess(), forward(), visualize() and
-        postprocess() according to the actual demands."""
-        results = [{}, {}, {}, {}]
-        dispatched_kwargs = set()
-
-        # Dispatch kwargs according to self.func_kwargs
-        for func_name, func_kwargs in self.func_kwargs.items():
-            for func_kwarg in func_kwargs:
-                if func_kwarg in kwargs:
-                    dispatched_kwargs.add(func_kwarg)
-                    results[self.func_order[func_name]][func_kwarg] = kwargs[
-                        func_kwarg]
-
-        return results
-
-    @torch.no_grad()
-    def __call__(self, **kwargs) -> Union[Dict, List[Dict]]:
-        """Call the inferencer.
-
-        Args:
-            kwargs: Keyword arguments for the inferencer.
-
-        Returns:
-            Union[Dict, List[Dict]]: Results of inference pipeline.
-        """
-        self._update_extra_parameters(**kwargs)
-
-        params = self._dispatch_kwargs(**kwargs)
-        preprocess_kwargs = self.base_params[0].copy()
-        preprocess_kwargs.update(params[0])
-        forward_kwargs = self.base_params[1].copy()
-        forward_kwargs.update(params[1])
-        visualize_kwargs = self.base_params[2].copy()
-        visualize_kwargs.update(params[2])
-        postprocess_kwargs = self.base_params[3].copy()
-        postprocess_kwargs.update(params[3])
-
-        data = self.preprocess(**preprocess_kwargs)
-        preds = self.forward(data, **forward_kwargs)
-        imgs = self.visualize(preds, **visualize_kwargs)
-        results = self.postprocess(preds, imgs, **postprocess_kwargs)
-        return results
-
-    def get_extra_parameters(self) -> List[str]:
-        """Each inferencer may has its own parameters. Call this function to
-        get these parameters.
-
-        Returns:
-            List[str]: List of unique parameters.
-        """
-        return list(self.extra_parameters.keys())
-
-    def postprocess(
-        self,
-        preds: PredType,
-        imgs: Optional[List[np.ndarray]] = None,
-        is_batch: bool = False,
-        get_datasample: bool = False,
-    ) -> Union[ResType, Tuple[ResType, np.ndarray]]:
-        """Postprocess predictions.
-
-        Args:
-            preds (List[Dict]): Predictions of the model.
-            imgs (Optional[np.ndarray]): Visualized predictions.
-            is_batch (bool): Whether the inputs are in a batch.
-                Defaults to False.
-            get_datasample (bool): Whether to use Datasample to store
-                inference results. If False, dict will be used.
-
-        Returns:
-            result (Dict): Inference results as a dict.
-            imgs (torch.Tensor): Image result of inference as a tensor or
-                tensor list.
-        """
-        results = preds
-        if not get_datasample:
-            results = []
-            for pred in preds:
-                result = self._pred2dict(pred)
-                results.append(result)
-        if not is_batch:
-            results = results[0]
-        return results, imgs
-
-    def _pred2dict(self, pred_tensor: torch.Tensor) -> Dict:
-        """Extract elements necessary to represent a prediction into a
-        dictionary. It's better to contain only basic data elements such as
-        strings and numbers in order to guarantee it's json-serializable.
-
-        Args:
-            pred_tensor (torch.Tensor): The tensor to be converted.
-
-        Returns:
-            dict: The output dictionary.
-        """
-        result = {}
-        result['infer_results'] = pred_tensor
-        return result
-
-    def visualize(self,
-                  inputs: list,
-                  preds: Any,
-                  show: bool = False,
-                  result_out_dir: str = '',
-                  **kwargs) -> List[np.ndarray]:
-        """Visualize predictions.
-
-        Customize your visualization by overriding this method. visualize
-        should return visualization results, which could be np.ndarray or any
-        other objects.
-
-        Args:
-            inputs (list): Inputs preprocessed by :meth:`_inputs_to_list`.
-            preds (Any): Predictions of the model.
-            show (bool): Whether to display the image in a popup window.
-                Defaults to False.
-            result_out_dir (str): Output directory of images. Defaults to ''.
-
-        Returns:
-            List[np.ndarray]: Visualization results.
-        """
-        results = (preds[:, [2, 1, 0]] + 1.) / 2.
-
-        # save images
-        if result_out_dir:
-            mkdir_or_exist(os.path.dirname(result_out_dir))
-            utils.save_image(results, result_out_dir)
-
-        return results
-
-
-def inferencer():
+def Inferencer(model_name: str = None,
+               model_setting: int = None,
+               model_config: str = None,
+               model_ckpt: str = None,
+               device: torch.device = None,
+               extra_parameters: Dict = None,
+               seed: int = 2022,
+               **kwargs) -> None:
     """MMEdit API for mmediting models inference.
 
     Args:
@@ -264,6 +59,7 @@ def inferencer():
 
         >>> # see demo/mmediting_inference_tutorial.ipynb for more examples
     """
+
     inference_supported_models = [
         # colorization models
         'inst_colorization',
@@ -274,7 +70,14 @@ def inferencer():
         'sagan',
 
         # unconditional models
+        'dcgan',
+        'wgan-gp',
+        'lsgan',
+        'ggan',
+        'pggan',
         'styleganv1',
+        'styleganv2',
+        'styleganv3',
 
         # matting models
         'gca',
@@ -307,204 +110,154 @@ def inferencer():
     inference_supported_models_cfg = {}
     inference_supported_models_cfg_inited = False
 
-    def __init__(self,
-                 model_name: str = None,
-                 model_setting: int = None,
-                 model_config: str = None,
-                 model_ckpt: str = None,
-                 device: torch.device = None,
-                 extra_parameters: Dict = None,
-                 seed: int = 2022,
-                 **kwargs) -> None:
-        init_default_scope('mmedit')
-        inferencer_kwargs = {}
-        inferencer_kwargs.update(
-            self._get_inferencer_kwargs(model_name, model_setting,
-                                        model_config, model_ckpt,
-                                        extra_parameters))
-        # self.inferencer = MMEditInferencer(
-        #     device=device, seed=seed, **inferencer_kwargs)
+    register_all_modules(init_default_scope=True)
+    MMEdit.init_inference_supported_models_cfg()
+    inferencer_kwargs = {}
+    inferencer_kwargs.update(
+        self._get_inferencer_kwargs(model_name, model_setting, model_config,
+                                    model_ckpt, extra_parameters))
+    self.inferencer = MMEditInferencer(
+        device=device, seed=seed, **inferencer_kwargs)
 
+    self.task = task
+    if self.task in ['conditional', 'Conditional GANs']:
+        self.inferencer = ConditionalInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['unconditional', 'Unconditional GANs']:
+        self.inferencer = UnconditionalInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['matting', 'Matting']:
+        self.inferencer = MattingInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['inpainting', 'Inpainting']:
+        self.inferencer = InpaintingInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['translation', 'Image2Image Translation']:
+        self.inferencer = TranslationInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['restoration', 'Image Super-Resolution']:
+        self.inferencer = RestorationInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['video_restoration', 'Video Super-Resolution']:
+        self.inferencer = VideoRestorationInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['video_interpolation', 'Video Interpolation']:
+        self.inferencer = VideoInterpolationInferencer(config, ckpt, device,
+                                                       extra_parameters)
+    elif self.task in ['text2image', 'Text2Image']:
+        self.inferencer = Text2ImageInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    elif self.task in ['3D_aware_generation', '3D-aware Generation']:
+        self.inferencer = EG3DInferencer(
+            config, ckpt, device, extra_parameters, seed=seed)
+    else:
+        raise ValueError(f'Unknown inferencer task: {self.task}')
+    """Get the kwargs for the inferencer."""
+    kwargs = {}
 
-#         self.task = task
-#         if self.task in ['conditional', 'Conditional GANs']:
-#             self.inferencer = ConditionalInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['colorization', 'Colorization']:
-#             self.inferencer = ColorizationInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['unconditional', 'Unconditional GANs']:
-#             self.inferencer = UnconditionalInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['matting', 'Matting']:
-#             self.inferencer = MattingInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['inpainting', 'Inpainting']:
-#             self.inferencer = InpaintingInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['translation', 'Image2Image']:
-#             self.inferencer = TranslationInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['restoration', 'Image Super-Resolution']:
-#             self.inferencer = RestorationInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['video_restoration', 'Video Super-Resolution']:
-#             self.inferencer = VideoRestorationInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['video_interpolation', 'Video Interpolation']:
-#             self.inferencer = VideoInterpolationInferencer(
-#                 config, ckpt, device, extra_parameters)
-#         elif self.task in ['text2image', 'Text2Image']:
-#             self.inferencer = Text2ImageInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         elif self.task in ['3D_aware_generation', '3D-aware Generation']:
-#             self.inferencer = EG3DInferencer(
-#                 config, ckpt, device, extra_parameters, seed=seed)
-#         else:
-#             raise ValueError(f'Unknown inferencer task: {self.task}')
+    if model_name is not None:
+        cfgs = self.get_model_config(model_name)
+        kwargs['task'] = cfgs['task']
+        setting_to_use = 0
+        if model_setting:
+            setting_to_use = model_setting
+        config_dir = cfgs['settings'][setting_to_use]['Config']
+        config_dir = config_dir[config_dir.find('configs'):]
+        kwargs['config'] = os.path.join(
+            osp.dirname(__file__), '..', config_dir)
+        kwargs['ckpt'] = cfgs['settings'][setting_to_use]['Weights']
 
-    inference_supported_models_cfg_inited = False
+    if model_config is not None:
+        if kwargs.get('config', None) is not None:
+            warnings.warn(
+                f'{model_name}\'s default config '
+                f'is overridden by {model_config}', UserWarning)
+        kwargs['config'] = model_config
 
-    def _get_inferencer_kwargs(self, model_name: Optional[str],
-                               model_setting: Optional[int],
-                               model_config: Optional[str],
-                               model_ckpt: Optional[str],
-                               extra_parameters: Optional[Dict]) -> Dict:
-        """Get the kwargs for the inferencer."""
-        kwargs = {}
+    if model_ckpt is not None:
+        if kwargs.get('ckpt', None) is not None:
+            warnings.warn(
+                f'{model_name}\'s default checkpoint '
+                f'is overridden by {model_ckpt}', UserWarning)
+        kwargs['ckpt'] = model_ckpt
 
-        if model_name is not None:
-            cfgs = self.get_model_config(model_name)
-            kwargs['task'] = cfgs['task']
-            setting_to_use = 0
-            if model_setting:
-                setting_to_use = model_setting
-            config_dir = cfgs['settings'][setting_to_use]['Config']
-            config_dir = config_dir[config_dir.find('configs'):]
-            kwargs['config'] = os.path.join(
-                osp.dirname(__file__), '..', config_dir)
-            kwargs['ckpt'] = cfgs['settings'][setting_to_use]['Weights']
-
-        if model_config is not None:
-            if kwargs.get('config', None) is not None:
-                warnings.warn(
-                    f'{model_name}\'s default config '
-                    f'is overridden by {model_config}', UserWarning)
-            kwargs['config'] = model_config
-
-        if model_ckpt is not None:
-            if kwargs.get('ckpt', None) is not None:
-                warnings.warn(
-                    f'{model_name}\'s default checkpoint '
-                    f'is overridden by {model_ckpt}', UserWarning)
-            kwargs['ckpt'] = model_ckpt
-
-        if extra_parameters is not None:
-            kwargs['extra_parameters'] = extra_parameters
+    if extra_parameters is not None:
+        kwargs['extra_parameters'] = extra_parameters
 
         return kwargs
 
-    def print_extra_parameters(self):
-        """Print the unique parameters of each kind of inferencer."""
-        extra_parameters = self.inferencer.get_extra_parameters()
-        print(extra_parameters)
 
-    def infer(self,
-              img: InputsType = None,
-              video: InputsType = None,
-              label: InputsType = None,
-              trimap: InputsType = None,
-              mask: InputsType = None,
-              result_out_dir: str = '',
-              **kwargs) -> Union[Dict, List[Dict]]:
-        """Infer edit model on an image(video).
+def print_extra_parameters(self):
+    """Print the unique parameters of each kind of inferencer."""
+    extra_parameters = self.inferencer.get_extra_parameters()
+    print(extra_parameters)
 
-        Args:
-            img (str): Img path.
-            video (str): Video path.
-            label (int): Label for conditional or unconditional models.
-            trimap (str): Trimap path for matting models.
-            mask (str): Mask path for inpainting models.
-            result_out_dir (str): Output directory of result image or video.
-                Defaults to ''.
 
-        Returns:
-            Dict or List[Dict]: Each dict contains the inference result of
-            each image or video.
-        """
-        return self.inferencer(
-            img=img,
-            video=video,
-            label=label,
-            trimap=trimap,
-            mask=mask,
-            result_out_dir=result_out_dir,
-            **kwargs)
+def get_model_config(self, model_name: str) -> Dict:
+    """Get the model configuration including model config and checkpoint url.
 
-    def get_model_config(self, model_name: str) -> Dict:
-        """Get the model configuration including model config and checkpoint
-        url.
+    Args:
+        model_name (str): Name of the model.
+    Returns:
+        dict: Model configuration.
+    """
+    if model_name not in self.inference_supported_models:
+        raise ValueError(f'Model {model_name} is not supported.')
+    else:
+        return self.inference_supported_models_cfg[model_name]
 
-        Args:
-            model_name (str): Name of the model.
-        Returns:
-            dict: Model configuration.
-        """
-        if model_name not in self.inference_supported_models:
-            raise ValueError(f'Model {model_name} is not supported.')
-        else:
-            return self.inference_supported_models_cfg[model_name]
 
-    @staticmethod
-    def init_inference_supported_models_cfg() -> None:
-        inference_supported_models_cfg_inited = False
-        if not inference_supported_models_cfg_inited:
-            all_cfgs_dir = osp.join(osp.dirname(__file__), '..', 'configs')
+@staticmethod
+def init_inference_supported_models_cfg() -> None:
+    if not MMEdit.inference_supported_models_cfg_inited:
+        all_cfgs_dir = osp.join(osp.dirname(__file__), '..', 'configs')
 
-            for model_name in inference_supported_models:
-                meta_file_dir = osp.join(all_cfgs_dir, model_name,
-                                         'metafile.yml')
-                with open(meta_file_dir, 'r') as stream:
-                    parsed_yaml = yaml.safe_load(stream)
-                task = parsed_yaml['Models'][0]['Results'][0]['Task']
-                inference_supported_models_cfg[model_name] = {}
-                inference_supported_models_cfg[model_name][
-                    'task'] = task  # noqa
-                inference_supported_models_cfg[model_name][
-                    'settings'] = parsed_yaml['Models']  # noqa
+        for model_name in MMEdit.inference_supported_models:
+            meta_file_dir = osp.join(all_cfgs_dir, model_name, 'metafile.yml')
+            with open(meta_file_dir, 'r') as stream:
+                parsed_yaml = yaml.safe_load(stream)
+            task = parsed_yaml['Models'][0]['Results'][0]['Task']
+            MMEdit.inference_supported_models_cfg[model_name] = {}
+            MMEdit.inference_supported_models_cfg[model_name][
+                'task'] = task  # noqa
+            MMEdit.inference_supported_models_cfg[model_name][
+                'settings'] = parsed_yaml['Models']  # noqa
 
-            inference_supported_models_cfg_inited = True
+        MMEdit.inference_supported_models_cfg_inited = True
 
-    @staticmethod
-    def get_inference_supported_models() -> List:
-        """static function for getting inference supported modes."""
-        return inference_supported_models
 
-    @staticmethod
-    def get_inference_supported_tasks() -> List:
-        """static function for getting inference supported tasks."""
-        inference_supported_models_cfg_inited = False
-        if not inference_supported_models_cfg_inited:
-            init_inference_supported_models_cfg()
+@staticmethod
+def get_inference_supported_models() -> List:
+    """static function for getting inference supported modes."""
+    return MMEdit.inference_supported_models
 
-        supported_task = set()
-        for key in inference_supported_models_cfg.keys():
-            if inference_supported_models_cfg[key]['task'] \
-               not in supported_task:
-                supported_task.add(inference_supported_models_cfg[key]['task'])
-        return list(supported_task)
 
-    @staticmethod
-    def get_task_supported_models(task: str) -> List:
-        """static function for getting task supported models."""
-        if not inference_supported_models_cfg_inited:
-            init_inference_supported_models_cfg()
+@staticmethod
+def get_inference_supported_tasks() -> List:
+    """static function for getting inference supported tasks."""
+    if not MMEdit.inference_supported_models_cfg_inited:
+        MMEdit.init_inference_supported_models_cfg()
 
-        supported_models = []
-        for key in inference_supported_models_cfg.keys():
-            if inference_supported_models_cfg[key]['task'] == task:
-                supported_models.append(key)
-        return supported_models
+    supported_task = set()
+    for key in MMEdit.inference_supported_models_cfg.keys():
+        if MMEdit.inference_supported_models_cfg[key]['task'] \
+            not in supported_task:
+            supported_task.add(
+                MMEdit.inference_supported_models_cfg[key]['task'])
+    return list(supported_task)
+
+
+@staticmethod
+def get_task_supported_models(task: str) -> List:
+    """static function for getting task supported models."""
+    if not MMEdit.inference_supported_models_cfg_inited:
+        MMEdit.init_inference_supported_models_cfg()
+
+    supported_models = []
+    for key in MMEdit.inference_supported_models_cfg.keys():
+        if MMEdit.inference_supported_models_cfg[key]['task'] == task:
+            supported_models.append(key)
+    return supported_models
 
 
 @torch.no_grad()
