+++ mmedit/utils/__init__.py
@@ -1,8 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .cli import modify_args
 from .img_utils import get_box_info, reorder_image, tensor2img, to_numpy
-from .io_utils import (MMEDIT_CACHE_DIR, download_from_url, read_frames,
-                       read_image)
+from .io_utils import (MMEDIT_CACHE_DIR, VIDEO_EXTENSIONS, download_from_url,
+                       read_frames, read_image)
 # TODO replace with engine's API
 from .logger import print_colored_log
 from .sampler import get_sampler
@@ -12,7 +12,8 @@ from .trans_utils import (add_gaussian_noise, adjust_gamma, bbox2mask,
                           brush_stroke_mask, calculate_grid_size,
                           get_irregular_mask, make_coord, pad_sequence,
                           random_bbox, random_choose_unknown)
-from .typing import ConfigType, ForwardInputs, LabelVar, NoiseVar, SampleList
+from .typing import (ConfigType, ForwardInputs, InputType, LabelVar, NoiseVar,
+                     SampleList)
 
 __all__ = [
     'modify_args', 'print_colored_log', 'register_all_modules', 'try_import',
@@ -21,6 +22,6 @@ __all__ = [
     'random_choose_unknown', 'add_gaussian_noise', 'adjust_gamma',
     'make_coord', 'bbox2mask', 'brush_stroke_mask', 'get_irregular_mask',
     'random_bbox', 'reorder_image', 'to_numpy', 'get_box_info', 'init_model',
-    'read_image', 'read_frames', 'set_random_seed', 'delete_cfg',
-    'calculate_grid_size', 'pad_sequence'
+    'read_image', 'read_frames', 'set_random_seed', 'delete_cfg', 'InputType',
+    'calculate_grid_size', 'pad_sequence', 'VIDEO_EXTENSIONS'
 ]
